Moving fb.h
